requirements.txt
streamlit==1.38.0
pdfplumber==0.11.0
python-docx==1.1.2
docx2txt==0.8
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
spacy==3.7.4